backend/Flask_Backend/requirements.txt
flask
flask-cors
python-dotenv
cloudinary
supabase
postgrest
pydantic
werkzeug